feat: notice 예외 처리 추가

+++ back/src/main/java/com/ssafy/soyu/notice/controller/NoticeController.java
@@ -1,15 +1,19 @@
 package com.ssafy.soyu.notice.controller;
 
+import static com.ssafy.soyu.util.response.ErrorResponseEntity.*;
+
 import com.ssafy.soyu.notice.service.NoticeService;
-import com.ssafy.soyu.util.fcm.service.FcmService;
+import com.ssafy.soyu.util.response.ErrorCode;
+import com.ssafy.soyu.util.response.exception.CustomException;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -19,22 +23,62 @@ import org.springframework.web.bind.annotation.RestController;
 @Tag(name = "Notice 컨트롤러", description = "Notice API 입니다.")
 public class NoticeController {
 
-  private final FcmService fcmService;
   private final NoticeService noticeService;
 
   @GetMapping("")
-  public ResponseEntity<?> selectNotice(@RequestAttribute Long memberId){
+  public ResponseEntity<?> selectNotice(HttpServletRequest request) {
+    Long memberId = (Long) request.getAttribute("memberId");
+    if (memberId == null) {
+      return toResponseEntity(ErrorCode.USER_NOT_FOUND);
+    }
+
     return new ResponseEntity<>(noticeService.findNotice(memberId), HttpStatus.OK);
   }
 
-  @DeleteMapping("")
-  public ResponseEntity<?> deleteNotice(@RequestAttribute Long noticeId){
+  /**
+   * 알림 삭제 처리<br/>
+   * 사용자 - 알림 매칭 여부 확인
+   * @param request 사용자 식별자 정보
+   * @param noticeId 확인할 알림의 식별자
+   */
+  @DeleteMapping("/{noticeId}")
+  public ResponseEntity<?> deleteNotice(HttpServletRequest request, @PathVariable Long noticeId) {
+    Long memberId = (Long) request.getAttribute("memberId");
+    if (memberId == null) {
+      return toResponseEntity(ErrorCode.USER_NOT_FOUND);
+    }
+
+    //알림 유저 매칭 여부 확인
+    try {
+      noticeService.checkNotice(memberId, noticeId);
+    } catch (CustomException e) {
+      return toResponseEntity(ErrorCode.NOT_MATCH_NOTICE);
+    }
+
     noticeService.deleteNotice(noticeId);
     return new ResponseEntity<>(HttpStatus.OK);
   }
 
-  @PutMapping("")
-  public ResponseEntity<?> readNotice(@RequestAttribute Long noticeId){
+  /**
+   * 알림 읽음 처리<br/>
+   * 사용자 - 알림 매칭 여부 확인
+   * @param request 사용자 식별자 정보
+   * @param noticeId 확인할 알림의 식별자
+   */
+  @PutMapping("/{noticeId}")
+  public ResponseEntity<?> readNotice(HttpServletRequest request, @PathVariable Long noticeId) {
+    Long memberId = (Long) request.getAttribute("memberId");
+    if (memberId == null) {
+      return toResponseEntity(ErrorCode.USER_NOT_FOUND);
+    }
+
+    //알림 유저 매칭 여부 확인
+    try {
+      noticeService.checkNotice(memberId, noticeId);
+    } catch (CustomException e) {
+      return toResponseEntity(ErrorCode.NOT_MATCH_NOTICE);
+    }
+
     noticeService.readNotice(noticeId);
     return new ResponseEntity<>(HttpStatus.OK);
   }
